Organize bitcoin indexer queries in indexer directory
