+++ packages/stores-bitcoin/src/queries/indexer/address-txs.ts
@@ -1,7 +1,7 @@
 import { ChainGetter, QuerySharedContext } from "@keplr-wallet/stores";
-import { ObservableBitcoinIndexerQuery } from "./bitcoin-indexer";
+import { ObservableBitcoinIndexerQuery } from "../bitcoin-indexer";
 import { makeObservable } from "mobx";
-import { BitcoinTx } from "./types";
+import { BitcoinTx } from "../types";
 
 export class ObservableQueryBitcoinAddressTxsImpl extends ObservableBitcoinIndexerQuery<
   BitcoinTx[]
